app/requirements.txt
Flask
waitress
python-dateutil
psutil
PyYAML
